MAINT: refactor dense_data into get_data
